feat(ui-results): add `MatrixGridSynthesis`
